+++ julia/src/02-lecture/OneHotEncode.jl
@@ -8,7 +8,7 @@ Convert class labels to one-hot vectors for classification.
 - `num_classes::Int`: Total number of classes
 
 # Returns
-- `Vector{Float64}`: One-hot encoded vector
+- `Vector{Float32}`: One-hot encoded vector
 
 # Example
 ```julia
@@ -16,7 +16,7 @@ one_hot_encode(3, 10)  # Returns [0, 0, 1, 0, 0, 0, 0, 0, 0, 0]
 ```
 """
 function one_hot_encode(label::Int, num_classes::Int)
-    encoded = zeros(Float64, num_classes)
+    encoded = zeros(Float32, num_classes)
     encoded[label] = 1.0
     return encoded
 end
